scripts/ShellcodeWrapper/requirements.txt
pycrypto
pyscrypt